v 1.3.0
Win32/Carberp method tweaked again to work from Windows 7 up to Windows
10 b10041.

+++ Source/Akagi/inject.c
@@ -4,9 +4,9 @@
 *
 *  TITLE:       INJECT.C
 *
-*  VERSION:     1.10
+*  VERSION:     1.30
 *
-*  DATE:        28 Mar 2015
+*  DATE:        30 Mar 2015
 *
 *  Inject module.
 *
@@ -62,7 +62,7 @@ BOOL ucmInjectExplorer(
 		//
 		// Allocate buffer in target process and write itself inside
 		//
-		remotebuffer = VirtualAllocEx(hProcess, NULL, opth->SizeOfImage,
+		remotebuffer = VirtualAllocEx(hProcess, NULL, (SIZE_T)opth->SizeOfImage,
 			MEM_COMMIT | MEM_RESERVE, PAGE_EXECUTE_READWRITE);
 
 		if (remotebuffer == NULL) {
